# openapi.yaml
openapi: "3.1.0"
info:
  title: LeanStock API
  version: "2.0.0"
  description: |
    Multi-tenant inventory management system with atomic transfers,
    dead stock decay, predictive reordering, reservation system,
    and full supplier/purchase-order workflows.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://leanstock.onrender.com
    description: Production (Render.com)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /products:
    get:
      tags: [Products]
      summary: List products (cursor pagination, category filter)
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              example:
                items: [{id: 1, name: "Widget Pro", sku: "WGT-001", price: 999.99, current_discount: 0.0, days_in_inventory: 0, min_stock: 10}]
                next_cursor: null

    post:
      tags: [Products]
      summary: Create product (manager+)
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: "Widget Pro"
              sku: "WGT-001"
              category: "electronics"
              price: 999.99
              min_stock: 10
      responses:
        "201":
          description: Product created
        "409":
          description: SKU already exists in tenant
  /products/{product_id}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
        "404":
          description: Not found
    put:
      tags: [Products]
      summary: Update product (manager+)
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: "Widget Pro v2"
              price: 1099.99
      responses:
        "200":
          description: Updated product
    delete:
      tags: [Products]
      summary: Soft-delete product (manager+)
      parameters:
        - name: product_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated

  # ── Locations ──────────────────────────────────────────────────────────────
...
  /purchase-orders:
    get:
      tags: [Purchase Orders]
      summary: List purchase orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, confirmed, received, cancelled]
      responses:
        "200":
          description: PO list with items
    post:
      tags: [Purchase Orders]
      summary: Create purchase order (draft, manager+)
      requestBody:
        required: true
        content:
          application/json:
            example:
              supplier_id: 1
              location_id: 1
              notes: "Urgent restock"
              items:
                - product_id: 1
                  quantity_ordered: 50
                  unit_price: 100.00
      responses:
        "201":
          description: PO created in draft
  /purchase-orders/{po_id}:
    get:
      tags: [Purchase Orders]
      summary: Get PO with items
      parameters:
        - name: po_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PO details
  /purchase-orders/{po_id}/send:
    post:
      tags: [Purchase Orders]
      summary: Send PO to supplier (draft → sent, sends email)
      parameters:
        - name: po_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PO sent, email queued
  /purchase-orders/{po_id}/receive:
    post:
      tags: [Purchase Orders]
      summary: Receive goods — updates inventory atomically
      parameters:
        - name: po_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            example:
              items:
                - product_id: 1
                  quantity_received: 50
      responses:
        "200":
          description: Received, inventory updated
  /purchase-orders/{po_id}/cancel:
    post:
      tags: [Purchase Orders]
      summary: Cancel PO
      parameters:
        - name: po_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled

  # ── Admin ──────────────────────────────────────────────────────────────────
...